Move src/env.ts to src/constants/env.ts
