+++ frontend/src/components/ApplicationProvider.tsx
@@ -9,7 +9,7 @@ import { ReactQueryDevtools } from 'react-query/devtools';
 import { Hydrate } from 'react-query/hydration';
 
 import { MediaContextProvider } from '@/components/media';
-import { PROD } from '@/env';
+import { PROD } from '@/constants/env';
 import { SearchStoreProvider } from '@/store/search/context';
 import { theme } from '@/theme';
