+++ frontend/src/pages/_app.tsx
@@ -17,9 +17,9 @@ import { ComponentType, ReactNode } from 'react';
 import { ApplicationProvider } from '@/components/ApplicationProvider';
 import { Layout } from '@/components/Layout';
 import { PageMetadata } from '@/components/PageMetadata';
+import { PROD } from '@/constants/env';
 import { DEFAULT_PLUGIN_DATA, DEFAULT_REPO_DATA } from '@/constants/plugin';
 import { LoadingStateProvider } from '@/context/loading';
-import { PROD } from '@/env';
 import { usePageTransitions } from '@/hooks';
 import SearchPage from '@/pages/index';
 import PluginPage from '@/pages/plugins/[name]';
